+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.2.2 (unreleased)
 
+- Added `destroyAll` function
 - Stopped refresh when chart is destroyed
 
 ## 3.2.1 (2020-07-23)